requirements.txt
streamlit==1.33.0
mediapipe==0.10.11
opencv-python==4.9.0.80
Pillow==10.3.0
numpy==1.26.4